+++ src/main/java/org/c4marathon/assignment/controller/TransferController.java
@@ -1,7 +1,7 @@
 package org.c4marathon.assignment.controller;
 
 import org.c4marathon.assignment.common.response.ApiResponse;
-import org.c4marathon.assignment.dto.account.TransferRequestDto;
+import org.c4marathon.assignment.dto.transfer.TransferRequestDto;
 import org.c4marathon.assignment.usecase.TransferUseCase;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
